+++ system/config.dev.php
@@ -1,0 +1,2 @@
+<?php
+return ['app'=>['debug'=>true]];